Give neovim virtual text a high deafult priority (#418)
This allows the suggestions to appear on top of coc.nvim inlay hints,
whereas with the default priority they appear below the inlay hint which
makes the suggestions difficult to read.

The default value is 65535, and the priority can be customized by setting the
g:codeium_virtual_text_priority and b:codeium_virtual_text_priority
variables.